+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -11,6 +11,7 @@ mod read_bytes;
 mod read_text;
 mod rm;
 mod rmdir;
+mod temp_file;
 mod touch;
 mod write_bytes;
 mod write_text;
@@ -37,6 +38,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(read_text::create(&package))?;
     commands.set(rm::create(&package))?;
     commands.set(rmdir::create(&package))?;
+    commands.set(temp_file::create(&package))?;
     commands.set(touch::create(&package))?;
     commands.set(write_bytes::create(&package))?;
     commands.set(write_text::create(&package))?;